requirements.txt
aiogram
python-dotenv
telethon
pydantic-settings
